+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `--verify-proof` option to `import-object` command.
 - Added support for conversation following.
+- Added `export-posts` command.
 
 ### Changed